requirements.txt
dialoguekit
ollama
python-dotenv
flask-cors
pillow
requests
spotipy
